rename package pkg/certdb to satellite/certdb
